Rename headers to better resemble class names
