+++ bridge/cxx/include/bhxx/BhBase.hpp
