+++ bridge/cxx/include/bhxx/BhInstruction.hpp
@@ -18,9 +18,8 @@ GNU Lesser General Public License along with Bohrium.
 If not, see <http://www.gnu.org/licenses/>.
 */
 #pragma once
-#include "bh_array.hpp"
+#include "BhArray.hpp"
 #include <bh_instruction.hpp>
-#include <type_traits>
 
 namespace bhxx {
